scheduler: save state of experiments
I spent a few days on this, and ultimately am abandoning this
approach to use lammps. I do not have it in me to debug 18 different
lammps installs on common clusters. It worked previously because
we could deploy the exact environment alongside lammps that
the container was built with (and thus it run). Trying to remove
that extra layer, and just run these on the base OS is a huge
undertaking that I have decided is too much anguish to do.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>